requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
psycopg[binary]==3.1.18
psycopg_pool==3.2.2
pydantic==2.9.2
python-dotenv==1.0.1
temporalio==1.9.0
nats-py==2.7.2
